+++ src/asio/AsioEventLoop.cpp
@@ -9,8 +9,9 @@
 namespace hical
 {
 
-	AsioEventLoop::AsioEventLoop()
-		: workGuard_(std::make_unique<boost::asio::executor_work_guard<boost::asio::io_context::executor_type>>(
+	AsioEventLoop::AsioEventLoop(int concurrencyHint)
+		: ioContext_(concurrencyHint)
+		, workGuard_(std::make_unique<boost::asio::executor_work_guard<boost::asio::io_context::executor_type>>(
 			  boost::asio::make_work_guard(ioContext_)))
 	{
 	}